refactor(agenda): extract items table
