+++ apps/web/src/components/agenda/agenda-page.tsx
@@ -9,33 +9,19 @@ import type { AgendaItem, AgendaItemPriority, AgendaItemStatus } from '@stitch/s
 import { AGENDA_ITEM_PRIORITIES, AGENDA_ITEM_STATUSES } from '@stitch/shared/agenda/types';
 
 import { AgendaItemDetailSheet } from '@/components/agenda/agenda-item-detail';
-import { PRIORITY_LABELS, PRIORITY_VARIANTS, STATUS_LABELS, STATUS_VARIANTS } from '@/components/agenda/constants';
-import { formatDateInTz, useUserTimezone } from '@/components/agenda/utils';
+import { AgendaItemsTable } from '@/components/agenda/agenda-items-table';
+import { PRIORITY_LABELS, STATUS_LABELS } from '@/components/agenda/constants';
+import { useUserTimezone } from '@/components/agenda/utils';
 import { Icon } from '@/components/primitives/icon';
 import { Stack } from '@/components/primitives/stack';
 import { Text } from '@/components/primitives/text';
 import { Button } from '@/components/ui/button';
-import { Calendar } from '@/components/ui/calendar';
-import { Checkbox } from '@/components/ui/checkbox';
 import { ConfirmDialog } from '@/components/ui/confirm-dialog';
 import { Dialog, DialogContent, DialogFooter, DialogHeader, DialogTitle } from '@/components/ui/dialog';
-import { Empty, EmptyDescription, EmptyMedia, EmptyTitle } from '@/components/ui/empty';
 import { Input } from '@/components/ui/input';
 import { MetricCard } from '@/components/ui/metric-card';
 import { Page, PageContent, PageDescription, PageHeader, PageHeaderContent, PageIcon } from '@/components/ui/page';
-import {
-  Pagination,
-  PaginationContent,
-  PaginationEllipsis,
-  PaginationItem,
-  PaginationLink,
-  PaginationNext,
-  PaginationPrevious,
-} from '@/components/ui/pagination';
-import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover';
 import { Select, SelectContent, SelectItem, SelectTrigger } from '@/components/ui/select';
-import { Table } from '@/components/ui/table';
-import { createAppColumnHelper, useAppTable } from '@/hooks/table-hook';
 import {
   agendaItemsQueryOptions,
   agendaListsQueryOptions,
@@ -45,13 +31,6 @@ import {
   useUpdateAgendaItem,
   useUpdateAgendaList,
 } from '@/lib/queries/agenda';
-const columnHelper = createAppColumnHelper<AgendaItem>();
-
-const agendaFilterColumns = columnHelper.columns([
-  columnHelper.accessor('status', {}),
-  columnHelper.accessor('priority', {}),
-]);
-
 export function AgendaPage({ listId }: { listId?: string }) {
   const navigate = useNavigate();
   const timeZone = useUserTimezone();
@@ -82,17 +61,6 @@ export function AgendaPage({ listId }: { listId?: string }) {
   const active = all.filter((i) => i.status !== 'done' && i.status !== 'cancelled');
   const completed = all.filter((i) => i.status === 'done' || i.status === 'cancelled');
   const items = [...active, ...completed];
-  const table = useAppTable({
-    data: items,
-    columns: agendaFilterColumns,
-    getRowId: (item) => item.id,
-    state: { columnFilters },
-    onColumnFiltersChange: (updater) => {
-      setColumnFilters(updater);
-      setPage(1);
-    },
-    manualFiltering: true,
-  });
   const totalPages = itemsData?.totalPages ?? 0;
   const total = itemsData?.total ?? 0;
 
@@ -137,6 +105,14 @@ export function AgendaPage({ listId }: { listId?: string }) {
     updateMutation.mutate({ id: itemId, updates: { dueAt } });
   }
 
+  function setFilter(id: 'status' | 'priority', value: string | null) {
+    setColumnFilters((current) => [
+      ...current.filter((filter) => filter.id !== id),
+      ...(!value || value === 'all' ? [] : [{ id, value }]),
+    ]);
+    setPage(1);
+  }
+
   function handleCreate() {
     if (!newTitle.trim()) return;
     createMutation.mutate(
@@ -177,21 +153,6 @@ export function AgendaPage({ listId }: { listId?: string }) {
   }
 
   const currentPage = (itemsData?.page ?? page) - 1;
-  let pageNumbers: number[];
-  if (totalPages <= 1) {
-    pageNumbers = [];
-  } else {
-    const firstPage = 0;
-    const lastPage = totalPages - 1;
-    const start = Math.max(firstPage, currentPage - 1);
-    const end = Math.min(lastPage, currentPage + 1);
-    const pages = new Set<number>([firstPage, lastPage]);
-    for (let index = start; index <= end; index += 1) {
-      pages.add(index);
-    }
-    pageNumbers = [...pages].toSorted((a, b) => a - b);
-  }
-
   const totalOpen = lists.reduce((sum, l) => sum + l.itemCounts.open, 0);
   const totalInProgress = lists.reduce((sum, l) => sum + l.itemCounts.in_progress, 0);
   const totalOverdue = lists.reduce((sum, l) => sum + l.itemCounts.overdue, 0);
@@ -277,9 +238,7 @@ export function AgendaPage({ listId }: { listId?: string }) {
 
         {/* Toolbar */}
         <Stack direction="row" wrap align="center" gap="m">
-          <Select
-            value={filterStatus ?? 'all'}
-            onValueChange={(value) => table.getColumn('status')?.setFilterValue(value === 'all' ? undefined : value)}>
+          <Select value={filterStatus ?? 'all'} onValueChange={(value) => setFilter('status', value)}>
             <SelectTrigger className="w-40 bg-background">
               <Text as="span" variant="body" truncate>
                 <Text as="span" variant="body" tone="muted">
@@ -298,9 +257,7 @@ export function AgendaPage({ listId }: { listId?: string }) {
             </SelectContent>
           </Select>
 
-          <Select
-            value={filterPriority ?? 'all'}
-            onValueChange={(value) => table.getColumn('priority')?.setFilterValue(value === 'all' ? undefined : value)}>
+          <Select value={filterPriority ?? 'all'} onValueChange={(value) => setFilter('priority', value)}>
             <SelectTrigger className="w-40 bg-background">
               <Text as="span" variant="body" truncate>
                 <Text as="span" variant="body" tone="muted">
@@ -320,118 +277,21 @@ export function AgendaPage({ listId }: { listId?: string }) {
           </Select>
         </Stack>
 
-        {/* Table */}
-        <Table.Container>
-          <Table.Scroller>
-            <Table.Root className="min-w-175 table-fixed">
-              <Table.Header>
-                <Table.Row className="hover:bg-transparent">
-                  <Table.Head className="w-10 text-center" />
-                  <Table.Head className="w-full min-w-0">Title</Table.Head>
-                  <Table.Head className="w-24 text-center">Status</Table.Head>
-                  <Table.Head className="w-20 text-center">Priority</Table.Head>
-                  {!listId && <Table.Head className="w-24 text-center">List</Table.Head>}
-                  <Table.Head className="w-28">Due</Table.Head>
-                  <Table.Head className="w-24" />
-                </Table.Row>
-              </Table.Header>
-              <Table.Body>
-                {isLoading ? (
-                  <Table.SkeletonRows
-                    columns={[
-                      { className: 'w-10' },
-                      { className: 'w-full max-w-0 min-w-0 overflow-hidden', skeletonClassName: 'h-4' },
-                      { className: 'w-24', skeletonClassName: 'mx-auto h-5 w-16 rounded-full' },
-                      { className: 'w-20', skeletonClassName: 'mx-auto h-5 w-14 rounded-full' },
-                      ...(!listId ? [{ className: 'w-24' }] : []),
-                      { className: 'w-28', skeletonClassName: 'h-4 w-16' },
-                      { className: 'w-24', skeletonClassName: 'ml-auto h-7 w-16 rounded-lg' },
-                    ]}
-                  />
-                ) : table.getRowModel().rows.length === 0 ? (
-                  <Table.EmptyRow colSpan={listId ? 6 : 7}>
-                    <Empty>
-                      <EmptyMedia variant="icon">
-                        <Icon as={ListTodoIcon} size="m" />
-                      </EmptyMedia>
-                      <EmptyTitle>No agenda items</EmptyTitle>
-                      <EmptyDescription>Create items from chat or click "New Item" to get started.</EmptyDescription>
-                    </Empty>
-                  </Table.EmptyRow>
-                ) : (
-                  <>
-                    {table.getRowModel().rows.map(({ original: item }) => (
-                      <AgendaItemRow
-                        key={item.id}
-                        item={item}
-                        showListColumn={!listId}
-                        timeZone={timeZone}
-                        deletePending={deleteMutation.isPending && deleteMutation.variables === item.id}
-                        onToggleDone={() => handleToggleDone(item)}
-                        onEdit={() => openItem(item)}
-                        onDelete={() => setItemToDelete(item)}
-                        onDateChange={handleDateChange}
-                      />
-                    ))}
-                  </>
-                )}
-              </Table.Body>
-            </Table.Root>
-          </Table.Scroller>
-
-          {totalPages > 1 ? (
-            <div className="border-t border-border px-space-l py-space-l">
-              <Pagination>
-                <PaginationContent>
-                  <PaginationItem>
-                    <PaginationPrevious
-                      href="#"
-                      onClick={(event) => {
-                        event.preventDefault();
-                        if (page > 1) setPage((c) => c - 1);
-                      }}
-                      className={page <= 1 ? 'pointer-events-none opacity-50' : undefined}
-                    />
-                  </PaginationItem>
-                  {pageNumbers.map((pageNumber, index) => {
-                    const previousPage = pageNumbers[index - 1];
-                    const showGap = pageNumber - previousPage > 1;
-                    return (
-                      <React.Fragment key={`page-${pageNumber}`}>
-                        {showGap ? (
-                          <PaginationItem>
-                            <PaginationEllipsis />
-                          </PaginationItem>
-                        ) : null}
-                        <PaginationItem>
-                          <PaginationLink
-                            href="#"
-                            isActive={pageNumber === currentPage}
-                            onClick={(event) => {
-                              event.preventDefault();
-                              setPage(pageNumber + 1);
-                            }}>
-                            {pageNumber + 1}
-                          </PaginationLink>
-                        </PaginationItem>
-                      </React.Fragment>
-                    );
-                  })}
-                  <PaginationItem>
-                    <PaginationNext
-                      href="#"
-                      onClick={(event) => {
-                        event.preventDefault();
-                        if (page < totalPages) setPage((c) => c + 1);
-                      }}
-                      className={page >= totalPages ? 'pointer-events-none opacity-50' : undefined}
-                    />
-                  </PaginationItem>
-                </PaginationContent>
-              </Pagination>
-            </div>
-          ) : null}
-        </Table.Container>
+        <AgendaItemsTable
+          items={items}
+          isLoading={isLoading}
+          showListColumn={!listId}
+          timeZone={timeZone}
+          page={page}
+          currentPage={currentPage}
+          totalPages={totalPages}
+          deletingItemId={deleteMutation.isPending ? deleteMutation.variables : undefined}
+          onPageChange={setPage}
+          onToggleDone={handleToggleDone}
+          onEdit={openItem}
+          onDelete={setItemToDelete}
+          onDateChange={handleDateChange}
+        />
       </PageContent>
 
       {/* Detail sheet */}
@@ -491,114 +351,5 @@ export function AgendaPage({ listId }: { listId?: string }) {
         isPending={deleteListMutation.isPending}
       />
     </Page>
-  );
-}
-
-type AgendaItemRowProps = {
-  item: AgendaItem;
-  showListColumn: boolean;
-  timeZone: string;
-  deletePending?: boolean;
-  onToggleDone: () => void;
-  onEdit: () => void;
-  onDelete: () => void;
-  onDateChange: (itemId: string, dueAt: number | null) => void;
-};
-
-function AgendaItemRow({
-  item,
-  showListColumn,
-  timeZone,
-  deletePending,
-  onToggleDone,
-  onEdit,
-  onDelete,
-  onDateChange,
-}: AgendaItemRowProps) {
-  const [dateOpen, setDateOpen] = React.useState(false);
-  const [nowMs] = React.useState(() => Date.now());
-  const isDone = item.status === 'done' || item.status === 'cancelled';
-  const isOverdue = item.dueAt && item.dueAt < nowMs && item.status !== 'done' && item.status !== 'cancelled';
-
-  return (
-    <Table.Row className={isDone ? 'opacity-50' : undefined}>
-      <Table.Cell className="w-10 text-center">
-        <Checkbox
-          checked={item.status === 'done'}
-          onCheckedChange={onToggleDone}
-          aria-label={item.status === 'done' ? 'Mark as open' : 'Mark as done'}
-        />
-      </Table.Cell>
-
-      <Table.Cell className="w-full max-w-0 min-w-0 overflow-hidden">
-        <Table.Title className={`block ${isDone ? 'text-muted-foreground line-through' : ''}`}>
-          {item.title}
-        </Table.Title>
-        {item.description && (
-          <Table.Text className={`block truncate ${isDone ? 'line-through' : ''}`}>{item.description}</Table.Text>
-        )}
-      </Table.Cell>
-
-      <Table.Cell className="w-24 text-center">
-        <Table.Badge variant={STATUS_VARIANTS[item.status]} size="xs">
-          {STATUS_LABELS[item.status]}
-        </Table.Badge>
-      </Table.Cell>
-
-      <Table.Cell className="w-20 text-center">
-        <Table.Badge variant={PRIORITY_VARIANTS[item.priority]} size="xs">
-          {PRIORITY_LABELS[item.priority]}
-        </Table.Badge>
-      </Table.Cell>
-
-      {showListColumn && (
-        <Table.Cell className="w-24 text-center text-xs text-muted-foreground">{item.listName ?? '—'}</Table.Cell>
-      )}
-
-      <Table.Cell className="w-28">
-        <Popover open={dateOpen} onOpenChange={setDateOpen}>
-          <PopoverTrigger
-            className={`inline-flex cursor-pointer rounded-sm px-space-xs py-space-2xs text-xs transition-colors hover:bg-muted ${isOverdue ? 'font-medium text-destructive' : 'text-muted-foreground'}`}>
-            {item.dueAt ? formatDateInTz(item.dueAt, timeZone) : '—'}
-          </PopoverTrigger>
-          <PopoverContent align="start" className="w-auto p-space-none">
-            <Calendar
-              mode="single"
-              selected={item.dueAt ? new Date(item.dueAt) : undefined}
-              onSelect={(date) => {
-                if (date) {
-                  const y = date.getFullYear();
-                  const m = date.getMonth();
-                  const d = date.getDate();
-                  const noon = new Date(y, m, d, 12, 0, 0);
-                  onDateChange(item.id, noon.getTime());
-                } else {
-                  onDateChange(item.id, null);
-                }
-                setDateOpen(false);
-              }}
-              defaultMonth={item.dueAt ? new Date(item.dueAt) : undefined}
-            />
-          </PopoverContent>
-        </Popover>
-      </Table.Cell>
-
-      <Table.Cell className="w-24">
-        <Table.Actions>
-          <Button type="button" variant="ghost" size="icon-sm" onClick={onEdit} aria-label={`Edit ${item.title}`}>
-            <Icon as={PencilIcon} size="s" />
-          </Button>
-          <Button
-            type="button"
-            variant="ghost"
-            size="icon-sm"
-            onClick={onDelete}
-            disabled={deletePending}
-            aria-label={`Delete ${item.title}`}>
-            <Icon as={Trash2Icon} size="s" tone="destructive" />
-          </Button>
-        </Table.Actions>
-      </Table.Cell>
-    </Table.Row>
   );
 }
